refactor(team-discussion): split pure view into TeamDiscussionPanelView, fix jsdom test import
The committed TeamDiscussionPanel.test.tsx imports the view through the
container file, whose transport/globalState import chain pulls in lottie-web —
which crashes at import time under jsdom (canvas API), so the suite could not
even collect the tests.

The presentational half now lives in TeamDiscussionPanelView.tsx with no heavy
imports (same split rationale as CaseHandoverCurtain); the container re-exports
it so stories and existing imports stay valid. Test imports the view file
directly — 7 view tests now actually run (17/17 in the module, full suite
919 green), tsc/eslint/stylelint clean, Storybook states verified in the built
bundle (open-with-messages, archived-read-only).

Refs OpenResilienceInitiative/ORISO-Frontend#514
